Added Text, EditableText, and Image bindings. Fixed Parent; added IndexInParent. Change CoordType to uint to match updated spec.
svn path=/trunk/at-spi-sharp/; revision=134539

+++ at-spi/Component.cs
@@ -116,7 +116,7 @@ namespace Atspi
 		LastDefined
 	}
 
-	public enum CoordType : short
+	public enum CoordType : uint
 	{
 		Screen,
 		Window
